Logic cleanup
